Module storm::prelude

Structs

AsyncOnceCell

Ctx

HashTable

OnceCell

A thread-safe cell which can be written to only once.

ProviderContainer

A dependency container to be able to instantiate and provide connection to databases.

QueueRwLock

VecTable

Traits

ApplyLog

Entity

Get

Insert

InsertMut

Remove

Tag

Transaction

Attribute Macros

indexing

Derive Macros

Ctx