Module storm::provider

Structs

LoadAllKeyOnly

LoadArgs

ProviderContainer

A dependency container to be able to instantiate and provide connection to databases.

TransactionProvider

Traits

Delete

LoadAll

LoadOne

Provider

ProviderFactory

Upsert

UpsertMut

This trait is implemented when the entity or the key must be changed while insert or update is performed. This case appears when there are triggers on the table, a sequence / identity column.