Trait storm::provider::UpsertMut

pub trait UpsertMut<E: Entity>: Send + Sync {
    fn upsert_mut<'a>(
        &'a self,
        k: &'a mut E::Key,
        v: &'a mut E
    ) -> BoxFuture<'a, Result<()>>;
}

This trait is implemented when the entity or the key must be changed while insert or update is performed. This case appears when there are triggers on the table, a sequence / identity column.

UpsertMut trait is not the same as the Upsert trait since the former need to take a key and entity as mutable parameters.

Required Methods

fn upsert_mut<'a>(
&'a self,
k: &'a mut E::Key,
v: &'a mut E
) -> BoxFuture<'a, Result<()>>