Struct storm::TblTransaction

pub struct TblTransaction<'a, 'b, E: Entity + EntityAccessor> { /* private fields */ }

Implementations

impl<'a, 'b, E> TblTransaction<'a, 'b, E>
where E: Entity + EntityAccessor, E::Key: Eq + Hash,

pub fn get(&self, k: &E::Key) -> Option<&E>
where Self: Get<E>,

gets a reference from the log or the underlying ctx.

You can take the TblTransaction by ownership and have a longer lifetime for the & by using the Self::into_ref method.

pub fn insert<'c>( &'c mut self, k: E::Key, v: E, track: &'c E::TrackCtx, ) -> BoxFuture<'c, Result<()>>
where Self: Insert<E>,

pub fn insert_all<'c, I>( &'c mut self, iter: I, track: &'c E::TrackCtx, ) -> BoxFuture<'c, Result<usize>>
where I: IntoIterator<Item = (E::Key, E)> + Send + 'a, I::IntoIter: Send, Self: Insert<E>,

pub fn insert_mut<'c>( &'c mut self, k: E::Key, v: E, track: &'c E::TrackCtx, ) -> BoxFuture<'c, Result<E::Key>>
where Self: InsertMut<E>,

pub fn insert_mut_all<'c, I>( &'c mut self, iter: I, track: &'c E::TrackCtx, ) -> BoxFuture<'c, Result<usize>>
where I: IntoIterator<Item = (E::Key, E)> + Send + 'a, I::IntoIter: Send, Self: InsertMut<E>,

pub fn into_ref(self, k: &E::Key) -> Option<&'b E>
where E: Entity + EntityAccessor + LogAccessor, E::Key: Eq + Hash, E::Tbl: Get<E>,

pub fn log(&self) -> Option<&FxHashMap<E::Key, LogState<E>>>
where E: Entity + EntityAccessor + LogAccessor,

pub fn remove<'c>( &'c mut self, k: E::Key, track: &'c E::TrackCtx, ) -> BoxFuture<'c, Result<()>>
where Self: Remove<E>,

pub fn remove_all<'c, K>( &'c mut self, keys: K, track: &'c E::TrackCtx, ) -> BoxFuture<'c, Result<usize>>
where Self: Remove<E>, K: IntoIterator<Item = E::Key> + Send + 'c, K::IntoIter: Send,

pub async fn remove_filter<F>( &mut self, filter: F, track: &E::TrackCtx, ) -> Result<()>
where E: EntityAccessor + LogAccessor, E::Key: Clone + Eq + Hash, F: FnMut(&E::Key, &E) -> bool, for<'c> &'c E::Tbl: IntoIterator<Item = (&'c E::Key, &'c E)> + Get<E>, Self: Remove<E>,

pub fn tbl(&self) -> &'a E::Tbl

pub async fn update_with<F>( &mut self, updater: F, track: &E::TrackCtx, ) -> Result<()>
where E: EntityAccessor + LogAccessor + ToOwned<Owned = E>, E::Key: Clone + Eq + Hash, F: for<'c> FnMut(&'c E::Key, &'c mut Cow<'_, E>), for<'c> &'c E::Tbl: IntoIterator<Item = (&'c E::Key, &'c E)> + Get<E>, Self: Insert<E>,

pub async fn update_mut_with<F>( &mut self, updater: F, track: &E::TrackCtx, ) -> Result<()>
where E: EntityAccessor + LogAccessor + ToOwned<Owned = E>, E::Key: Clone + Eq + Hash, F: for<'c> FnMut(&'c E::Key, &'c mut Cow<'_, E>), for<'c> &'c E::Tbl: IntoIterator<Item = (&'c E::Key, &'c E)> + Get<E>, Self: InsertMut<E>,