Trait storm::IsDefined

pub trait IsDefined {
    // Required method
    fn is_defined(&self) -> bool;
}

Indicate if a key is not zero. This trait must be implemented on keys used by the UpsertMut when a key is flagged as Identity (MS SQL). In such a case, the provider will check that property to determine if the entity must be inserted or updated.

Required Methods

fn is_defined(&self) -> bool