Trait storm::OneToManyFromIter

pub trait OneToManyFromIter<ONE, MANY>: IntoIterator<Item = (ONE, MANY)> + Sizedwhere
    ONE: Clone + From<usize> + Into<usize>,{
    fn collect_sort<F>(self) -> OneToMany<ONE, MANY>
    where
        MANY: Ord,
    { ... }
    fn collect_sort_by<F>(self, cmp: F) -> OneToMany<ONE, MANY>
    where
        F: Fn(&MANY, &MANY) -> Ordering,
    { ... }
    fn collect_sort_by_key<F, K>(self, key_cmp: F) -> OneToMany<ONE, MANY>
    where
        F: Fn(&MANY) -> K,
        K: Ord,
    { ... }
    fn collect_sort_dedup(self) -> OneToMany<ONE, MANY>
    where
        MANY: Ord,
    { ... }
    fn collect_sort_dedup_by<F>(self, cmp: F) -> OneToMany<ONE, MANY>
    where
        F: Fn(&MANY, &MANY) -> Ordering,
    { ... }
    fn collect_sort_dedup_by_key<F, K>(self, key_cmp: F) -> OneToMany<ONE, MANY>
    where
        F: Fn(&MANY) -> K,
        K: Ord,
    { ... }
}

Provided Methods

fn collect_sort<F>(self) -> OneToMany<ONE, MANY>where
MANY: Ord,

Collect the items and sort them.

fn collect_sort_by<F>(self, cmp: F) -> OneToMany<ONE, MANY>where
F: Fn(&MANY, &MANY) -> Ordering,

Collect the items and sort them by the cmp function specified.

fn collect_sort_by_key<F, K>(self, key_cmp: F) -> OneToMany<ONE, MANY>where
F: Fn(&MANY) -> K,
K: Ord,

Collect the items and sort them using a key_cmp function.

fn collect_sort_dedup(self) -> OneToMany<ONE, MANY>where
MANY: Ord,

fn collect_sort_dedup_by<F>(self, cmp: F) -> OneToMany<ONE, MANY>where
F: Fn(&MANY, &MANY) -> Ordering,

Collect the items and sort them by the cmp function specified. Also removes duplicate MANY items using the same cmp function.

fn collect_sort_dedup_by_key<F, K>(self, key_cmp: F) -> OneToMany<ONE, MANY>where
F: Fn(&MANY) -> K,
K: Ord,

Collect the items and sort them by the key_cmp function specified. Also removes duplicate MANY items using the same key_cmp function.