Crate storm

Re-exports

pub use len::macro_check_max_len;

pub use provider::ProviderContainer;

pub use trx_iter::TrxIter;

pub use async_cell_lock;

pub use attached;

pub use cache;

pub use parking_lot;

pub use tokio;

pub use vec_map;

pub use version_tag;

pub use gc::*;

Modules

gc

mem

prelude

provider

trx_iter

vars

Macros

tri

version_deps

Structs

AsyncOnceCell

Ctx

CtxLocks

CtxTransaction

HashTable

Logs

OnChange

OnChanged

OnRemove

OnceCell

A thread-safe cell which can be written to only once.

OneToMany

QueueRwLock

TblTransaction

VecMap

VecTable

VersionTag

Enums

Changed

Error

FieldsOrStr

LogState

Constants

EV_CREATED

OBJ_INDEX

OBJ_TABLE

Traits

Accessor

ApplyLog

AsRefAsync

AsRefOpt

AsyncTryFrom

ChangeHandler

ChangedHandler

CommitHandler

CtxTypeInfo

Entity

EntityAccessor

EntityFields

EntityOf

EntityValidate

Fields

FromRefOpt

Get

GetMut

Init

Insert

InsertIfChanged

InsertMut

InsertMutIfChanged

IsDefined

Indicate if a key is not zero. This trait must be implemented on keys used by the UpsertMut when a key is flagged as Identity (MS SQL). In such a case, the provider will check that property to determine if the entity must be inserted or updated.

IteratorExt

Len

LogAccessor

NotifyTag

OneToManyFromIter

Remove

RemoveHandler

Tag

Transaction

Functions

change_depth

Returns a stack depth level 1.. when run inside the on_change event. Each time the event is nested, the depth increase. When called oustide, will returns 0.

register_on_commit_handler

Type Aliases

BoxFuture

Deps

Log

LogVar

LogsVar

Result

TblVar

Vars

Attribute Macros

indexing

Derive Macros

Ctx

LocksAwait

MssqlDelete

MssqlLoad

MssqlSave

NoopDelete

NoopLoad

NoopSave