Crate storm

Re-exports

pub use provider::ProviderContainer;

pub use registry::set_date_provider;

pub use trx_iter::TrxIter;

pub use async_cell_lock;

pub use cache;

pub use extobj;

pub use fast_set;

pub use fxhash;

pub use linkme;

pub use parking_lot;

pub use tokio;

pub use vec_map;

pub use version_tag;

pub use gc::*;

Modules

gc

indexing

mem

prelude

provider

registry

trx_iter

Macros

flat_set_adapt

hash_flat_set_adapt

one_adapt

single_set_adapt

tree_index_adapt

tri

version_deps

Structs

AppliedEvent

AsyncOnceCell

ChangedEvent

ClearEvent

CommitEvent

Ctx

CtxExt

CtxLocks

CtxTransaction

GcEvent

HashTable

IntSet

Logs

OnceCell

A thread-safe cell which can be written to only once.

OneToMany

QueueRwLock

RemovedEvent

RemovingEvent

TblTransaction

TblTransactionIter

TouchedEvent

UpsertedEvent

UpsertingEvent

VecMap

VecTable

VersionTag

Enums

ApplyOrder

Error

FieldsOrStr

Constants

EV_CREATED

OBJ_INDEX

OBJ_TABLE

Traits

ApplyLog

AsRefAsync

AsRefOpt

AsyncTryFrom

Clearable

CtxTypeInfo

Entity

EntityAccessor

EntityFields

EntityOf

EntityRemove

EntityUpsert

EntityUpsertMut

EntityValidate

Fields

FromRefOpt

Get

GetMut

IsDefined

Indicate if a key is not zero. This trait must be implemented on keys used by the UpsertMut when a key is flagged as Identity (MS SQL). In such a case, the provider will check that property to determine if the entity must be inserted or updated.

IteratorExt

Len

LogOf

NotifyTag

OneToManyFromIter

RefIntoIterator

Tag

Touchable

Transaction

Type Aliases

BoxFuture

CtxExtObj

CtxVar

Deps

Result

Attribute Macros

flat_set_index

hash_flat_set_index

indexing

one_index

register

single_set

tree_index

Derive Macros

Ctx

LocksAwait

MssqlDelete

MssqlLoad

MssqlSave

NoopDelete

NoopLoad

NoopSave