Crate storm

Re-exports

pub use async_cell_lock;

pub use attached;

pub use cache;

pub use gc::*;

pub use parking_lot;

pub use tokio;

pub use vec_map;

pub use version_tag;

Modules

gc

mem

prelude

provider

vars

Macros

version_deps

Structs

AsyncOnceCell

Ctx

CtxLocks

CtxTransaction

HashTable

OnceCell

A thread-safe cell which can be written to only once.

OneToMany

ProviderContainer

A dependency container to be able to instantiate and provide connection to databases.

QueueRwLock

TblTransaction

VecMap

VecTable

VersionTag

Enums

Error

FieldsOrStr

LogState

Constants

EV_CREATED

OBJ_INDEX

OBJ_TABLE

Traits

Accessor

ApplyLog

AsRefAsync

AsRefOpt

AsyncTryFrom

CtxTypeInfo

Entity

EntityAccessor

EntityFields

EntityOf

FromRefOpt

Get

GetMut

Init

Insert

InsertMut

IsDefined

Indicate if a key is not zero. This trait must be implemented on keys used by the UpsertMut when a key is flagged as Identity (MS SQL). In such a case, the provider will check that property to determine if the entity must be inserted or updated.

LogAccessor

NotifyTag

OneToManyFromIter

Remove

Tag

Transaction

Type Definitions

BoxFuture

Deps

Log

LogVar

Logs

Result

TblVar

Vars

Attribute Macros

indexing

Derive Macros

Ctx

LocksAwait

MssqlDelete

MssqlLoad

MssqlSave

NoopDelete

NoopLoad

NoopSave