Struct storm::VecMap

pub struct VecMap<K, V> { /* private fields */ }

Implementations

impl<K, V> VecMap<K, V>

pub fn par_iter(&self) -> ParIter<'_, K, V>
where K: Sync, V: Sync,

pub fn par_iter_mut(&mut self) -> ParIterMut<'_, K, V>
where K: Send, V: Send,

impl<K, V> VecMap<K, V>

pub const fn new() -> VecMap<K, V>

pub fn with_capacity(capacity: usize) -> VecMap<K, V>

pub fn clear(&mut self)

pub fn contains_key(&self, key: &K) -> bool
where K: Copy + Into<usize>,

pub fn entry(&mut self, key: K) -> Entry<'_, K, V>
where K: Copy + Into<usize>,

pub fn get(&self, key: &K) -> Option<&V>
where K: Copy + Into<usize>,

pub fn get_mut(&mut self, key: &K) -> Option<&mut V>
where K: Copy + Into<usize>,

pub fn insert(&mut self, key: K, value: V) -> Option<V>
where K: Copy + Into<usize>,

pub fn is_empty(&self) -> bool

pub fn iter(&self) -> Iter<'_, K, V>

pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

pub fn keys(&self) -> Keys<'_, K, V>

pub fn len(&self) -> usize

pub fn remove(&mut self, key: &K) -> Option<V>
where K: Copy + Into<usize>,

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&K, &V) -> bool, K: Copy + From<usize>,

Retains only the elements specified by the predicate.

In other words, remove all elements e such that f(&e) returns false. This method operates in place, visiting each element exactly once in the original order, and preserves the order of the retained elements.

Example

use vec_map::VecMap;

let mut map = VecMap::new();
map.insert(1usize, 10);
map.insert(2usize, 11);

map.retain(|_k, v| v > &10);

assert_eq!(map.len(), 1);
assert_eq!(map.into_iter().collect::<Vec<_>>(), vec![(2, 11)]);

pub fn shrink_to_fit(&mut self)

pub fn values(&self) -> Values<'_, K, V>

pub fn values_mut(&mut self) -> ValuesMut<'_, K, V>